requirements.txt
google-api-python-client
google-auth
gspread
requests
python-dotenv
oauthlib
requests-oauthlib
schedule
